refactor(cosh-ng): [shell] split large files
Review follow-ups on the large-file policy: extract the approval
trust domain (session trusted commands + run batch consent) from
runtime/state.rs into runtime/trust_state.rs (1013 -> 998 lines,
below merge base), and the model-independent action-row rendering
from ui/agent_render/approval.rs into approval_actions.rs
(1087 -> 944, below the main baseline); large-file inventory entries
updated accordingly.
