+++ src/cosh-ng/crates/cosh-shell/src/agent/poll.rs
@@ -407,7 +407,7 @@ fn poll_active_agent_run_with_policy<W: Write>(
             active_run.status_animation.clear(output)?;
         }
         // Turn-scope batch consent never outlives its run (issue #1773).
-        state.control.clear_run_batch_consent();
+        state.control.trust.clear_run_batch_consent();
         render_fresh_turn_recovery_notice(state, output)?;
         // Fresh-turn recovery is an internal fallback continuation.
         start_agent_run_with_origin(
