+++ src/cosh-ng/crates/cosh-shell/src/runtime/cancel.rs
@@ -117,7 +117,7 @@ fn cancel_active_agent_run<W: Write>(
         return Ok(());
     };
     // Turn-scope batch consent never outlives its run (issue #1773).
-    state.control.clear_run_batch_consent();
+    state.control.trust.clear_run_batch_consent();
 
     let cancellation_details_id = state.provider_cancellation_artifacts.record_cancelled_run(
         active_run.request.id.clone(),
